AP_Compass: solve typecast misbehaviour observed on one compiler version

+++ libraries/AP_Compass/CompassCalibrator.cpp
@@ -683,7 +683,7 @@ void CompassCalibrator::run_ellipsoid_fit()
 //////////// CompassSample public interface //////////////
 //////////////////////////////////////////////////////////
 
-#define COMPASS_CAL_SAMPLE_SCALE_TO_FIXED(__X) ((int16_t)constrain_float(roundf(__X*8.0f), INT16_MIN, INT16_MAX))
+#define COMPASS_CAL_SAMPLE_SCALE_TO_FIXED(__X) ((int16_t)constrain_float(roundf(__X*8.0f), float(int16_t(INT16_MIN)), float(int16_t(INT16_MAX))))
 #define COMPASS_CAL_SAMPLE_SCALE_TO_FLOAT(__X) (__X/8.0f)
 
 Vector3f CompassCalibrator::CompassSample::get() const {
